PROTON-1294: c++ examples: remove un-necessary 64-bit integers
64 bit integers cause portability problems on some platforms (e.g. need extra
atomic library support on ppc 32 bit) and do not enhance the examples, keep them simple.

+++ examples/cpp/broker.cpp
@@ -147,7 +147,7 @@ class queues {
   protected:
     typedef std::map<std::string, queue *> queue_map;
     queue_map queues_;
-    uint64_t next_id_; // Use to generate unique queue IDs.
+    int next_id_; // Use to generate unique queue IDs.
 };
 
 // A handler to implement broker logic
