+++ examples/cpp/direct_recv.cpp
@@ -37,8 +37,8 @@ class direct_recv : public proton::messaging_handler {
   private:
     std::string url;
     proton::listener listener;
-    uint64_t expected;
-    uint64_t received;
+    int expected;
+    int received;
 
   public:
     direct_recv(const std::string &s, int c) : url(s), expected(c), received(0) {}
@@ -49,7 +49,7 @@ class direct_recv : public proton::messaging_handler {
     }
 
     void on_message(proton::delivery &d, proton::message &msg) OVERRIDE {
-        if (proton::coerce<uint64_t>(msg.id()) < received) {
+        if (proton::coerce<int>(msg.id()) < received) {
             return; // Ignore duplicate
         }
